+++ src/base/window_animator.h
@@ -127,7 +127,7 @@ typedef window_animator_t* (*window_animator_create_t)(bool_t open, object_t* ar
  */
 struct _window_animator_t {
   uint32_t duration;
-  uint32_t start_time;
+  uint64_t start_time;
   easing_func_t easing;
 
   widget_t* prev_win;
@@ -153,11 +153,11 @@ struct _window_animator_t {
  * 更新动画。
  * @param {window_animator_t*} wa 窗口动画对象。
  * @param {canvas_t*} canvas 画布对象。
- * @param {uint32_t} time_ms 当前时间(毫秒)。
+ * @param {uint64_t} time_ms 当前时间(毫秒)。
  *
  * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
  */
-ret_t window_animator_update(window_animator_t* wa, uint32_t time_ms);
+ret_t window_animator_update(window_animator_t* wa, uint64_t time_ms);
 
 /**
  * @method window_animator_begin_frame
